Using rocket state for storing the server does not work
Rocket takes ownership of the managed value, but does not provide static
lifetimes for it (as the rocket ignite funtion can return). So we generally
cannot use it for storing the server while using a separate thread for the
hyper server. This may work when using a scoped thread library like
crossbeam.